backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.0
python-dotenv==1.0.0
PyMuPDF==1.23.8
python-docx==1.1.0
pandas==2.1.4
openpyxl==3.1.2
openai==1.92.0
redis==5.0.1 